fix: move types to separate folder and export
