+++ src/utilities.ts
@@ -7,14 +7,18 @@ import {
 } from "@stacks/wallet-sdk";
 import type {
   AddressNonces,
+  ContractSourceResponse,
   Transaction,
 } from "@stacks/stacks-blockchain-api-types";
 import { StackingClient } from "@stacks/stacking";
 import { TxBroadcastResult } from "@stacks/transactions";
-
-// define types of networks we allow
-// matches string definitions in Stacks.js
-export type NetworkType = "mainnet" | "testnet" | "devnet" | "mocknet";
+import {
+  AppConfig,
+  NamesDataResponse,
+  NetworkType,
+  POXResponse,
+  TransactionResponse,
+} from "./types";
 
 // validate network value
 export function validateNetwork(network: string | undefined): NetworkType {
@@ -93,13 +97,6 @@ export async function getFaucetDrop(
   return data;
 }
 
-// define structure of app config
-interface AppConfig {
-  NETWORK: NetworkType;
-  MNEMONIC: string;
-  ACCOUNT_INDEX: number;
-}
-
 // define default values for app config
 const DEFAULT_CONFIG: AppConfig = {
   NETWORK: "testnet",
@@ -111,7 +108,6 @@ const DEFAULT_CONFIG: AppConfig = {
 function loadConfig(): AppConfig {
   // Bun loads .env automatically
   // so nothing to load here first
-
   return {
     NETWORK: validateNetwork(process.env.NETWORK),
     MNEMONIC: process.env.MNEMONIC || DEFAULT_CONFIG.MNEMONIC,
@@ -272,18 +268,6 @@ export async function getNamesOwnedByAddress(network: string, address: string) {
   return data.names;
 }
 
-type NamesDataResponse = {
-  address: string;
-  blockchain: string;
-  expire_block?: number;
-  grace_period?: number;
-  last_txid?: string;
-  resolver?: string;
-  status?: string;
-  zonefile?: string;
-  zonefile_hash?: string;
-};
-
 // gets address by name from the hiro api
 export async function getAddressByName(network: string, name: string) {
   const apiUrl = getApiUrl(network);
@@ -319,67 +303,6 @@ export async function getNextNonce(network: string, address: string) {
   return nextNonce;
 }
 
-type Epoch = {
-  epoch_id: string;
-  start_height: number;
-  end_height: number;
-  block_limit: {
-    write_length: number;
-    write_count: number;
-    read_length: number;
-    read_count: number;
-    runtime: number;
-  };
-  network_epoch: number;
-};
-
-type RewardCycle = {
-  id: number;
-  min_threshold_ustx: number;
-  stacked_ustx: number;
-  is_pox_active: boolean;
-};
-
-type NextCycle = {
-  id: number;
-  min_threshold_ustx: number;
-  min_increment_ustx: number;
-  stacked_ustx: number;
-  prepare_phase_start_block_height: number;
-  blocks_until_prepare_phase: number;
-  reward_phase_start_block_height: number;
-  blocks_until_reward_phase: number;
-  ustx_until_pox_rejection: number | null;
-};
-
-type ContractVersion = {
-  contract_id: string;
-  activation_burnchain_block_height: number;
-  first_reward_cycle_id: number;
-};
-
-type POXResponse = {
-  contract_id: string;
-  pox_activation_threshold_ustx: number;
-  first_burnchain_block_height: number;
-  current_burnchain_block_height: number;
-  prepare_phase_block_length: number;
-  reward_phase_block_length: number;
-  reward_slots: number;
-  rejection_fraction: number | null;
-  total_liquid_supply_ustx: number;
-  current_cycle: RewardCycle;
-  next_cycle: NextCycle;
-  epochs: Epoch[];
-  min_amount_ustx: number;
-  prepare_cycle_length: number;
-  reward_cycle_id: number;
-  reward_cycle_length: number;
-  rejection_votes_left_required: number | null;
-  next_reward_cycle_in: number;
-  contract_versions: ContractVersion[];
-};
-
 export async function getPOXDetails(network: NetworkType) {
   const apiUrl = getApiUrl(network);
   const response = await fetch(`${apiUrl}/v2/pox`, {
@@ -392,12 +315,6 @@ export async function getPOXDetails(network: NetworkType) {
   }
   const data = (await response.json()) as POXResponse;
   return data;
-}
-
-interface ContractSourceResponse {
-  source: string;
-  publish_height: number;
-  proof: string;
 }
 
 export async function getContractSource(
@@ -453,105 +370,6 @@ export async function getAddressBalanceDetailed(
   } catch (error: any) {
     throw new Error(`Failed to get address balance: ${error.message}`);
   }
-}
-
-interface TransactionResponse {
-  limit: number;
-  offset: number;
-  total: number;
-  results: Array<{
-    tx: {
-      tx_id: string;
-      nonce: number;
-      fee_rate: string;
-      sender_address: string;
-      sponsor_nonce: number;
-      sponsored: boolean;
-      sponsor_address: string;
-      post_condition_mode: string;
-      post_conditions: Array<{
-        principal: {
-          type_id: string;
-        };
-        condition_code: string;
-        amount: string;
-        type: string;
-      }>;
-      anchor_mode: string;
-      block_hash: string;
-      block_height: number;
-      block_time: number;
-      block_time_iso: string;
-      burn_block_height: number;
-      burn_block_time: number;
-      burn_block_time_iso: string;
-      parent_burn_block_time: number;
-      parent_burn_block_time_iso: string;
-      canonical: boolean;
-      tx_index: number;
-      tx_status: string;
-      tx_result: {
-        hex: string;
-        repr: string;
-      };
-      event_count: number;
-      parent_block_hash: string;
-      is_unanchored: boolean;
-      microblock_hash: string;
-      microblock_sequence: number;
-      microblock_canonical: boolean;
-      execution_cost_read_count: number;
-      execution_cost_read_length: number;
-      execution_cost_runtime: number;
-      execution_cost_write_count: number;
-      execution_cost_write_length: number;
-      events: Array<{
-        event_index: number;
-        event_type: string;
-        tx_id: string;
-        contract_log: {
-          contract_id: string;
-          topic: string;
-          value: {
-            hex: string;
-            repr: string;
-          };
-        };
-      }>;
-      tx_type: string;
-      contract_call: {
-        contract_id: string;
-        function_name: string;
-      };
-      smart_contract: {
-        contract_id: string;
-      };
-      token_transfer: {
-        recipient_address: string;
-        amount: string;
-        memo: string;
-      };
-    };
-    stx_sent: string;
-    stx_received: string;
-    events: {
-      stx: {
-        transfer: number;
-        mint: number;
-        burn: number;
-      };
-      ft: {
-        transfer: number;
-        mint: number;
-        burn: number;
-      };
-      nft: {
-        transfer: number;
-        mint: number;
-        burn: number;
-      };
-    };
-  }>;
 }
 
 export async function getTransactionsByAddress(
